Move method to vcfutils
